+++ packages/@expo/cli/src/export/exportStaticAsync.ts
@@ -14,7 +14,7 @@ import {
   isScreenRouteNode,
   type RouteNode,
 } from 'expo-router/build/Route';
-import { getContextKey, stripGroupSegmentsFromPath } from 'expo-router/build/matchers';
+import { getRoutePathname, stripGroupSegmentsFromPath } from 'expo-router/build/matchers';
 import { shouldLinkExternally } from 'expo-router/build/utils/url';
 import type { PageHeaderInfo, RoutesManifest } from 'expo-server/private';
 import path from 'path';
@@ -292,7 +292,7 @@ export async function exportFromServerAsync(
         const data = await loaderResponse.json();
         // Transforms a `route.contextKey` into a normalized path. For example,
         // `./nested/[id]/index.tsx` becomes `/nested/[id]/index`
-        const loaderKey = getContextKey(route.contextKey);
+        const loaderKey = getRoutePathname(route.contextKey);
         const fileSystemPath = `_expo/loaders${loaderKey}`;
         files.set(fileSystemPath, {
           contents: JSON.stringify(data, null, 2),