+++ packages/expo-router/src/fork/getPathFromState.ts
@@ -337,15 +337,3 @@ const createNormalizedConfigs = (
       return [name, result];
     })
   );
-
-export function appendBaseUrl(
-  path: string,
-  baseUrl: string | undefined = process.env.EXPO_BASE_URL
-) {
-  if (process.env.NODE_ENV !== 'development') {
-    if (baseUrl) {
-      return `/${baseUrl.replace(/^\/+/, '').replace(/\/$/, '')}${path}`;
-    }
-  }
-  return path;
-}